Cleaned imports and externalized useConfig hook
